# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maximizing Your Chat with AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /retention:
    get:
      operationId: getRetention
      tags: [retention]
      summary: The tenant's data-retention policy and the plan cap
      description: >-
        Returns the caller tenant's chosen retention periods (days; null =
        unlimited) plus the maximum the active plan allows. Team members resolve
        to their owner.
      responses:
        "200":
          description: Retention policy and plan cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionInfo"
    put:
      operationId: updateRetention
      tags: [retention]
      summary: Update the tenant's data-retention policy
      description: >-
        Sets the retention periods (days; null = unlimited). Each value is
        clamped server-side to the active plan's retention cap. Super admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetentionInput"
      responses:
        "200":
          description: Updated retention policy and plan cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionInfo"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /credentials/{id}/contacts/status:
    get:
      operationId: getCredentialContactsStatus
      tags: [credentials]
      summary: Count of Google Contacts stored for the signed-in user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Contacts status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactsStatus" }
...
  /statuses:
    get:
      operationId: listStatuses
      tags: [whatsapp]
      summary: List WhatsApp statuses (mine + from contacts, last 24h)
      responses:
        "200":
          description: Status feed grouped by author
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsappStatusAuthor"
    post:
      operationId: postStatus
      tags: [whatsapp]
      summary: Post a text status (story) from the connected account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostStatusInput"
      responses:
        "200":
          description: Status posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappStatus2"
  /statuses/{id}:
    delete:
      operationId: deleteStatus
      tags: [whatsapp]
      summary: Delete one of my statuses (local row only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    WhatsappStatus2:
      type: object
      required:
        [id, statusType, textContent, backgroundColor, mediaUrl, caption,
         isMine, authorName, postedAt, expiresAt]
      properties:
        id:
          type: integer
        statusType:
          type: string
          enum: [text, image, video]
        textContent:
          type: ["string", "null"]
        backgroundColor:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        isMine:
          type: boolean
        authorName:
          type: string
        postedAt:
          type: string
        expiresAt:
          type: string
    WhatsappStatusAuthor:
      type: object
      required: [authorJid, authorName, authorPhone, profilePicUrl, isMine, statuses]
      properties:
        authorJid:
          type: string
        authorName:
          type: string
        authorPhone:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        isMine:
          type: boolean
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/WhatsappStatus2"
...
    RetentionInput:
      type: object
      description: >-
        Retention periods in days. null = unlimited (keep forever). Omitted
        fields are treated as null.
      properties:
        chatDays:
          type: ["integer", "null"]
          minimum: 1
        mediaDays:
          type: ["integer", "null"]
          minimum: 1
        logDays:
          type: ["integer", "null"]
          minimum: 1
        analyticsDays:
          type: ["integer", "null"]
          minimum: 1
    RetentionInfo:
      type: object
      required: [chatDays, mediaDays, logDays, analyticsDays, planLimitDays]
      properties:
        chatDays:
          type: ["integer", "null"]
          description: Max age (days) for chat messages; null = unlimited.
        mediaDays:
          type: ["integer", "null"]
          description: Max age (days) for stored media; null = unlimited.
        logDays:
          type: ["integer", "null"]
          description: Max age (days) for AI usage logs; null = unlimited.
        analyticsDays:
          type: ["integer", "null"]
          description: Max age (days) for usage/analytics snapshots; null = unlimited.
        planLimitDays:
          type: ["integer", "null"]
          description: Plan-imposed cap on selectable retention (days); null = no cap.
...
    PostStatusInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 700
        backgroundColor:
          type: string
          description: Hex color string e.g. "#0f3a4d" — defaults to WA green if omitted.
...
    ContactsStatus:
      type: object
      required: [count]
      properties:
        count: { type: integer }
...