# lib/api-zod/src/generated/types/retentionInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Maximizing Your Chat with AI API
 * OpenAPI spec version: 0.1.0
 */

export interface RetentionInfo {
  /**
     * Max age (days) for chat messages; null = unlimited.
     * @nullable
     */
  chatDays: number | null;
  /**
     * Max age (days) for stored media; null = unlimited.
     * @nullable
     */
  mediaDays: number | null;
  /**
     * Max age (days) for AI usage logs; null = unlimited.
     * @nullable
     */
  logDays: number | null;
  /**
     * Max age (days) for usage/analytics snapshots; null = unlimited.
     * @nullable
     */
  analyticsDays: number | null;
  /**
     * Plan-imposed cap on selectable retention (days); null = no cap.
     * @nullable
     */
  planLimitDays: number | null;
}