# lib/api-zod/src/generated/types/retentionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Maximizing Your Chat with AI API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Retention periods in days. null = unlimited (keep forever). Omitted fields are treated as null.
 */
export interface RetentionInput {
  /**
     * @minimum 1
     * @nullable
     */
  chatDays?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  mediaDays?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  logDays?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  analyticsDays?: number | null;
}